Finish to refactoring of Model
All logic has been moved to the Objects rather than in the ParseModel. The objects now control themselves and what we need to do with them. Start to work on refactoring of how selection of region will work. Will used a containerView that can move with PanGesture as well as on a button tap. Also added movement of the view on the loginView so that when the keyboard shows, all of the elements should be visible no matter the size of the screen.