Most of the change of flow
Changed so that RegionsList will be opened by tapping on a barButtonItem on the MapVC. Still need to add in logic so that we always have a default Region. Also addition of a convenience initializer for Directions so Directions will initiate with what they need just by passing in the NSDictionary.